Add a license (fixes #34)
I absolutely agree, that the license for wiki should be very permissive,
on the other hand I want to make sure that my work will be acknowledged
(somehow at least). I would like people to recontribute to my project,
but I don't want to force them to do so, so I decided to settle for a
three clause BSD license (which is also what is used for flask). I hope
that this is a valid choice for everyone.